Photogrammetry Project Commit
All the base models that were generated, as well as previews to said models and any later modified version (with the exception of 5 of the models for file size reasons) are all being posted here. I cannot upload the Meshroom Caches, as those totalled to literally 50+ GB when I can only upload a maximum of 2 GB to GitHub.